% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AbundancePlot.R
\name{AbundancePlot}
\alias{AbundancePlot}
\title{Plot cell abundances across groups in a Seurat object}
\usage{
AbundancePlot(
  object,
  group.by,
  split.by,
  replicate.by,
  draw_paths = FALSE,
  perform.stat.test = TRUE,
  test.choice = "t.test",
  ncols = 3,
  paired = FALSE,
  set_minimum_to_zero = FALSE,
  fill_colors = NULL,
  point_colors = NULL,
  p.adjust.method = "BH",
  filter.p.above = 1,
  step.increase = 0.15,
  title_size = 10,
  p_val_size = 2.5,
  plot_type = "box",
  bees = TRUE,
  x_lab = NULL,
  y_lab = NULL,
  point_size = 1,
  same_y_limit = FALSE,
  selected.groups = NULL,
  rotated_axis = FALSE,
  print_results = FALSE
)
}
\arguments{
\item{object}{Seurat object}

\item{group.by}{Which meta.data slot to use for grouping the cells}

\item{split.by}{Which meta.data slot to use for splitting the groups}

\item{replicate.by}{Which meta.data slot to use as split.by replicates}

\item{draw_paths}{Whether or not to draw paths in the abundance plots between
replicates if they exist}

\item{perform.stat.test}{Whether or not to perform a statistical test between
the splits}

\item{test.choice}{Which choice of test to use (supports t.test of wilcox)}

\item{ncols}{Number of columns to plot the groups as}

\item{paired}{Whether the t.test should be paired if using t.test}

\item{set_minimum_to_zero}{Whether to set minimum abundance on plots to 0
globally}

\item{point_colors}{Vector of colors to color points as}

\item{p.adjust.method}{Which method to use for p-value adjustment for mulitple
testing}

\item{filter.p.above}{Plot p-values only below this filter}

\item{step.increase}{Tweaking the space between brackets for p-values vertically}

\item{title_size}{Text size of each group plot's title}

\item{p_val_size}{Text size for p-values}

\item{plot_type}{Either box or violin plot}

\item{x_lab}{The label on the x axis}

\item{y_lab}{The label on the y axis}

\item{point_size}{Size of the points plotted}

\item{same_y_limit}{Whether to keep the upper y limit constant across all
groups.}

\item{selected.groups}{Optional. If not NULL, will only plot the groups
specified in this argument.}

\item{rotated_axis}{Whether to apply Seurat::RotatedAxis to the plots}

\item{print_results}{Prints to screen the results of the abundance calculations
and abundance statistical testing.}

\item{sina_shift}{Whether to use ggforce::geom_sina to pretty-shift the points}
}
\value{
Returns a patchwork plot of separate ggplots per group.by variable
}
\description{
Plot cell abundances across groups in a Seurat object
}
\details{
Plots the abundances of specific groups in the Seurat object
across a split.by variable, using replicate.by as replicates for each
split.by condition
}
