% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BackgatePlot.R
\name{BackgatePlot}
\alias{BackgatePlot}
\title{Plot a metadata selection in a biaxial FeatureScatter plot}
\usage{
BackgatePlot(
  object,
  feature1,
  feature2,
  metadata.col,
  metadata.selection,
  assay = "RNA",
  layer = "data",
  label.cols = c("red", "grey"),
  pt.size = 1,
  add.density = FALSE,
  density.col = "black",
  noise.zero = FALSE,
  noise.zero.bound = -0.01,
  raster = TRUE,
  raster.dpi = c(512, 512)
)
}
\arguments{
\item{object}{Seurat object}

\item{feature1}{The x-axis feature}

\item{feature2}{The y-axis feature}

\item{metadata.col}{The column within the object meta.data to select from}

\item{metadata.selection}{The entry within metadata.col to highlight}

\item{assay}{Which assay}

\item{layer}{Which layer in the assay, typically data}

\item{label.cols}{2-d vector specifying colors for labelling}

\item{pt.size}{Size of geom_point}

\item{add.density}{Whether to add geom_density_2d estimation to figure}

\item{density.col}{The color of the density contour (if added)}

\item{noise.zero}{Whether to simulate values from 0 to a negative value,
done for visual purposes ONLY and for kernel density estimation when
overlaying a density visualization. Will simulate values using a normal
distribution centered at 0, take absolute values, and then scale from
noise.zero.bound to 0.}

\item{noise.zero.bound}{The negative value to which to simulate zeros to.}

\item{high.p}{Which percentile to set as 1}
}
\value{
Returns a Seurat object with scaled values in the data layer of the
specified assay
}
\description{
Plot a metadata selection in a biaxial FeatureScatter plot
}
\details{
Scales the expression values for each feature in the specified assay
from 0 to 1, with an optional split.by argument
}
