% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildSNNfromKNN.R
\name{BuildSNNfromKNN}
\alias{BuildSNNfromKNN}
\title{Derive an SNN from a kNN Graph object in Seurat}
\usage{
BuildSNNfromKNN(
  object,
  knn_choice = "RNA_nn",
  snn_name = paste0(knn_choice, "_snn"),
  prune_snn = 1/15
)
}
\arguments{
\item{object}{Seurat object}

\item{knn_choice}{Name of Graph slot to use for SNN calculation}

\item{snn_name}{Name of Graph slot to store SNN}

\item{prune_snn}{Threshold at which to prune SNN (set to 0 below this)}
}
\value{
Returns a Seurat object with the SNN Graph stored.
}
\description{
Derive an SNN from a kNN Graph object in Seurat
}
\details{
code is adapted from @immunogenomics/singlecellmethods. Builds a
shared nearest-neighbors graph from an input kNN using matrix multiplication
so pretty quick.
}
\references{
code is adapted from @immunogenomics/singlecellmethods
}
