% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClipDSB.R
\name{ClipDSB}
\alias{ClipDSB}
\title{Clip DSB assay with quantiles per feature}
\usage{
ClipDSB(
  object,
  assay = "DSB",
  slot = "data",
  split.by = NULL,
  low.p = NULL,
  high.p = 0.999,
  new.assay = "sDSB"
)
}
\arguments{
\item{object}{Seurat object}

\item{assay}{Which assay, typically DSB}

\item{slot}{Which slot in the assay, typically data after DSB correction}

\item{split.by}{If not NULL, which metadata feature to split by  before
scaling.}

\item{low.p}{Which percentile below which to clip per feature}

\item{high.p}{Which percentile above which to clip per feature}

\item{new.assay}{Name of new assay to save values to}
}
\value{
Returns a Seurat object with clipped values in the data slot of the
specified assay
}
\description{
Clip DSB assay with quantiles per feature
}
\details{
Clips the expression values for each feature in the specified assay
from low.p to high.p, with an optional split.by argument.
}
