% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeuratPlotEnhancements.R
\name{DimPlotEdges}
\alias{DimPlotEdges}
\title{Seurat DimPlot enhancement}
\usage{
DimPlotEdges(
  object,
  graph.choice = "RNA_nn",
  reduction = "umap",
  line.thickness = 0.05,
  line.color = "grey",
  raster = FALSE,
  ...
)
}
\arguments{
\item{object}{Seurat object}

\item{graph.choice}{Which Graph to utilize for plotting with the DimPlot}

\item{reduction}{Which reduction to use in the DimPlot}

\item{line.thickness}{Thickness of lines connecting individual points in the
DimPlot based on the graph.choice Graph}

\item{line.color}{Color of the lines connecting individual points}

\item{raster}{Raseterization boolean}

\item{...}{Other arguments to Seurat::DimPlot}
}
\value{
Returns a DimPlot
}
\description{
Seurat DimPlot enhancement
}
\details{
Identical to DimPlot with the exception that edges can be drawn on
the DimPlot between cells connected in the specified Graph.
}
