% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeuratPlotEnhancements.R
\name{DimPlotPlus}
\alias{DimPlotPlus}
\title{Seurat DimPlot enhancement II}
\usage{
DimPlotPlus(
  reduction = "umap",
  shape.choice = 21,
  pt.size = 1.75,
  st.size = 0.25,
  ...
)
}
\arguments{
\item{reduction}{Which reduction to use in the DimPlot}

\item{shape.choice}{Which shape to use for the points int he DimPlot}

\item{pt.size}{Size of DimPlot points}

\item{st.size}{Size of DimPlot stencil around points}

\item{...}{Other arguments to Seurat::DimPlot}

\item{object}{Seurat object}
}
\value{
Returns a DimPlot
}
\description{
Seurat DimPlot enhancement II
}
\details{
Identical to DimPlot with the exception that points are drawn as
fill-able circles with a black border.
}
