% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dissociation.genes}
\alias{dissociation.genes}
\title{Dissociation-associated gene signature}
\format{
## `dissociation.genes`
A vector with 137 gene names:
\describe{
  \item{gene name}{gene name}
  ...
}
}
\source{
<https://www.nature.com/articles/nmeth.4437>
}
\usage{
dissociation.genes
}
\description{
The human gene names of the dissociation-associated gene signatures from the
"Single-cell sequencing reveals dissociation-induced gene expression in
tissue subpopulations." van den Brink SC, et al. Nat Methods 2017.
}
\keyword{datasets}
