% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeuratPlotEnhancements.R
\name{DoClusteredDotPlot}
\alias{DoClusteredDotPlot}
\title{Seurat DotPlot enhancement with hierarchical clustering}
\usage{
DoClusteredDotPlot(
  object,
  features = NULL,
  assay = "RNA",
  slot = "data",
  group.by = "seurat_clusters",
  max_zed = 3,
  y_text_size = 10,
  scale_rows = TRUE,
  plot_dendro = FALSE,
  auc_choice = 0.6,
  dot.scale = 5,
  cluster_cols = TRUE
)
}
\arguments{
\item{object}{Seurat object}

\item{features}{Which features to plot}

\item{assay}{Which assay to get expression values from}

\item{slot}{Which slot to pull the assay data from}

\item{group.by}{Which metadata slot to group cells by}

\item{max_zed}{Cutoff for maximum z-scores (absolute value)}

\item{y_text_size}{Size of text on y axis (features)}

\item{scale_rows}{Whether to scale the expression per row}

\item{plot_dendro}{Whether to plot accompanying x-axis and y-axis dendrograms}

\item{dot.scale}{Value by which to scale the dots in DotPlot}

\item{cluster_cols}{Whether to cluster the columns as well as the rows}
}
\value{
Returns a patchwork object of a DotPlot and dendrogram if specified
}
\description{
Seurat DotPlot enhancement with hierarchical clustering
}
\details{
Identical to DotPlot with the exception that the genes/features are
clustered using hierarchical clustering for prettier visualization
}
