% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heatmaps.R
\name{DoClusteredHeatmap}
\alias{DoClusteredHeatmap}
\title{Heatmap of average expression values per group}
\usage{
DoClusteredHeatmap(
  object,
  features = NULL,
  assay = "RNA",
  layer = "data",
  group.by = "seurat_clusters",
  max_zed = 3,
  y_text_size = 10,
  scale_rows = TRUE,
  plot_dendro = FALSE,
  dot.scale = 5,
  cluster_cols = TRUE,
  cluster_rows = TRUE,
  viridis_option = "B",
  viridis_direction = 1
)
}
\arguments{
\item{object}{Seurat object}

\item{features}{Which features to plot}

\item{assay}{Which assay to get expression values from}

\item{layer}{Which layer to pull the assay data from}

\item{group.by}{Which metadata layer to group cells by}

\item{max_zed}{Cutoff for maximum z-scores (absolute value)}

\item{y_text_size}{Size of text on y axis (features)}

\item{scale_rows}{Whether to scale the expression per row}

\item{plot_dendro}{Whether to plot accompanying x-axis and y-axis dendrograms}

\item{dot.scale}{Value by which to scale the dots in DotPlot}

\item{cluster_cols}{Whether to cluster the columns}

\item{cluster_rows}{Whether to cluster the rows}

\item{viridis_option}{The viridis color scale to use}

\item{viridis_direction}{The direction of the viridis color scale}
}
\value{
Returns a patchwork object of a heatmap and dendrogram if specified
}
\description{
Heatmap of average expression values per group
}
\details{
Plots the average expression value of the specified genes/features
for each group in the Seurat object, optionally using hierarchical
clustering for prettier visualization
}
