% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heatmaps.R
\name{DoGenesetHeatmap}
\alias{DoGenesetHeatmap}
\title{Heatmap of average expression values per group}
\usage{
DoGenesetHeatmap(
  seurat_object,
  assay = "AUC",
  slot = "counts",
  group.by = "seurat_clusters",
  diff_exp_results,
  max_zed = 3,
  plot_rownames = FALSE,
  label_features = NULL,
  y_text_size = 10,
  scale_rows = TRUE,
  p_val_choice = 0.01,
  logFC = 0.4,
  plot_all = FALSE,
  plot_dendro = FALSE,
  auc_choice = 0.6
)
}
\arguments{
\item{assay}{Which assay to get expression values from}

\item{slot}{Which slot to pull the assay data from}

\item{group.by}{Which metadata slot to group cells by}

\item{max_zed}{Cutoff for maximum z-scores (absolute value)}

\item{y_text_size}{Size of text on y axis (features)}

\item{scale_rows}{Whether to scale the expression per row}

\item{plot_dendro}{Whether to plot accompanying x-axis and y-axis dendrograms}

\item{object}{Seurat object}

\item{features}{Which features to plot}

\item{dot.scale}{Value by which to scale the dots in DotPlot}

\item{cluster_cols}{Whether to cluster the columns}

\item{cluster_rows}{Whether to cluster the rows}
}
\value{
Returns a patchwork object of a heatmap and dendrogram if specified
}
\description{
Heatmap of average expression values per group
}
\details{
Plots the average expression value of the specified genes/features
for each group in the Seurat object, optionally using hierarchical
clustering for prettier visualization
}
