% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractClusterProportions.R
\name{ExtractClusterProportions}
\alias{ExtractClusterProportions}
\title{Extract cell proportions across groups in a Seurat object}
\usage{
ExtractClusterProportions(object, group.by, split.by, replicate.by)
}
\arguments{
\item{object}{Seurat object}

\item{group.by}{Which meta.data slot to use for grouping the cells}

\item{split.by}{Which meta.data slot to use for splitting the groups}

\item{replicate.by}{Which meta.data slot to use as split.by replicates}
}
\value{
Returns a data frame of group.by abundances per replicate in the
split.by variable
}
\description{
Extract cell proportions across groups in a Seurat object
}
\details{
Returns the abundances of specific groups in the Seurat object
across a split.by variable, using replicate.by as replicates for each
split.by condition
}
