% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeuratPlotEnhancements.R
\name{FeaturePlotPlus}
\alias{FeaturePlotPlus}
\title{Seurat FeaturePlot enhancement}
\usage{
FeaturePlotPlus(
  object,
  feature,
  min.value = 0,
  assay = "RNA",
  viridis_choice = "D",
  ...
)
}
\arguments{
\item{object}{Seurat object}

\item{feature}{Which feature to plot (singular)}

\item{min.value}{Which threshold value to set as not-expressed/grey in the
FeaturePlot}

\item{assay}{Which assay the feature is located in}

\item{viridis_choice}{Which viridis palette to choose}

\item{...}{Other arguments to Seurat::FeaturePlot}
}
\value{
Returns a FeaturePlot
}
\description{
Seurat FeaturePlot enhancement
}
\details{
Identical to FeaturePlot with the exception that 0-values are set to
grey, which replicates the Monocle-type UMAP plots.
}
