% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonSeurat_utils.R
\name{FindJaccard}
\alias{FindJaccard}
\title{Determine overlap measure between a list of genes}
\usage{
FindJaccard(
  genelist,
  distance.measure = "jaccard",
  as.distance = FALSE,
  return.tidy = TRUE
)
}
\arguments{
\item{genelist}{Named list of gene vectors}

\item{distance.measure}{Which sort of distance measure to use (Jaccard by default)}

\item{as.distance}{Whether to transform similarities to distances}

\item{return.tidy}{Whether to return a tidy dataframe}
}
\value{
Returns a matrix of genelist distance measures
}
\description{
Determine overlap measure between a list of genes
}
