% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaDataPlot.R
\name{MetaDataPlot}
\alias{MetaDataPlot}
\title{Plot Seurat metadata distribution as a bar graph for indicated groups.}
\usage{
MetaDataPlot(
  object,
  group.by,
  split.by,
  as.freq = TRUE,
  label.count = FALSE,
  text.size = 3,
  angled.text = TRUE,
  label.angle = 45
)
}
\arguments{
\item{object}{Seurat object}

\item{group.by}{Which meta.data slot to use for assessment}

\item{split.by}{Which meta.data slot to use for group splitting}

\item{as.freq}{Whether to plot the group.by metadata as a frequency or
proportion}

\item{label.count}{Whether to add count on top of bar graph}

\item{angled.text}{Whether to angle the text to 45 degrees}
}
\value{
Returns a metadata plot
}
\description{
Plot Seurat metadata distribution as a bar graph for indicated groups.
}
\details{
Plots the group.by metadata as blocks on the y axis split across the
specified metadata in the split.by argument.
}
