% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonSeurat_utils.R
\name{ReadGMT}
\alias{ReadGMT}
\title{Read in a GMT file}
\usage{
ReadGMT(file_path = "", file_sep = ",")
}
\arguments{
\item{file_path}{path to .gmt file}

\item{file_sep}{which delimiter is used to separate entries in .gmt file}
}
\value{
Returns a data frame of the .gmt file terms and genes
}
\description{
Read in a GMT file
}
\references{
code adapted from clusterProfiler::read.gmt
}
