% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunHarmony2.R
\name{RunHarmony2}
\alias{RunHarmony2}
\title{Run Harmony with necessary internal object tinkering for Symphony}
\usage{
RunHarmony2(
  object,
  group.by.vars,
  reduction = "pca",
  dims.use = NULL,
  theta = NULL,
  lambda = NULL,
  sigma = 0.1,
  nclust = NULL,
  tau = 0,
  block.size = 0.05,
  max.iter.harmony = 10,
  max.iter.cluster = 20,
  epsilon.cluster = 1e-05,
  epsilon.harmony = 1e-04,
  plot_convergence = FALSE,
  verbose = TRUE,
  reference_values = NULL,
  reduction.save = "harmony",
  assay.use = "RNA",
  project.dim = TRUE,
  ...
)
}
\value{
Runs Harmony in the implementation of Symphony such that the Harmony
model is stored within the Seurat object for later use with Symphony, see
https://github.com/immunogenomics/symphony/blob/main/vignettes/utils_seurat.R
}
\description{
Run Harmony with necessary internal object tinkering for Symphony
}
\references{
copy-pasted from @immunogenomics/symphony github repo
}
