% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunHarmony2.R
\name{RunHarmonyForSymphony}
\alias{RunHarmonyForSymphony}
\title{Run Harmony with necessary internal object tinkering for Symphony}
\usage{
RunHarmonyForSymphony(
  object,
  group.by.vars,
  reduction.use = "pca",
  dims.use = NULL,
  reduction.save = "harmony",
  project.dim = TRUE,
  ...
)
}
\value{
Runs Harmony in the implementation of Symphony such that the Harmony
model is stored within the Seurat object for later use with Symphony, see
https://github.com/immunogenomics/symphony/blob/main/vignettes/utils_seurat.R
}
\description{
Run Harmony with necessary internal object tinkering for Symphony
}
\references{
copy-pasted from @immunogenomics/harmony github repo
}
