% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunSCA.R
\name{RunSCA}
\alias{RunSCA}
\title{Perform Shannon Component Analysis on a Seurat object}
\usage{
RunSCA(
  object,
  assay = "RNA",
  slot = "data",
  n.comps = 50L,
  iters = 3L,
  reduction.name = "sca",
  reduction.key = "sca_"
)
}
\arguments{
\item{object}{Seurat object}

\item{assay}{Which assay to use for SCA calculation}

\item{slot}{Which slot in the assay to use for SCA calculation}

\item{n.comps}{Number of SCs to calculate}

\item{iters}{How many iterations to perform for SCA}

\item{reduction.name}{Name of the Reduction object to store the SCA as}

\item{reduction.key}{Key preceding the numbers of SCA components in
reduction.name}
}
\value{
Returns a Seurat object with an SCA calculated
}
\description{
Perform Shannon Component Analysis on a Seurat object
}
\references{
Demeo and Berger, _biorxiv_ 2021. shannonca.readthedocs.io
}
