% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScaleADT.R
\name{ScaleADT}
\alias{ScaleADT}
\title{Scale ADT assay from 0 to 1 per feature}
\usage{
ScaleADT(
  object,
  assay = "ADT",
  slot = "data",
  split.by = NULL,
  low.p = 0.001,
  high.p = 0.999
)
}
\arguments{
\item{object}{Seurat object}

\item{assay}{Which assay, typically ADT}

\item{slot}{Which slot in the assay, typically data after CLR}

\item{split.by}{If not NULL, which metadata feature to split by  before
scaling.}

\item{low.p}{Which percentile to set as 0}

\item{high.p}{Which percentile to set as 1}
}
\value{
Returns a Seurat object with scaled values in the data slot of the
specified assay
}
\description{
Scale ADT assay from 0 to 1 per feature
}
\details{
Scales the expression values for each feature in the specified assay
from 0 to 1, with an optional split.by argument
}
