% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScaleDSB.R
\name{ScaleDSB}
\alias{ScaleDSB}
\title{Scale DSB assay to [0-1] per feature}
\usage{
ScaleDSB(
  object,
  assay = "DSB",
  layer = "data",
  split.by = NULL,
  high.p = 0.999,
  scale.to.one = FALSE,
  new.assay = "sDSB"
)
}
\arguments{
\item{object}{Seurat object}

\item{assay}{Which assay, typically DSB}

\item{layer}{Which layer in the assay, typically data after DSB correction}

\item{split.by}{If not NULL, which metadata feature to split by  before
scaling.}

\item{high.p}{Which percentile to set as the global 1 per feature}

\item{scale.to.one}{Whether to scale the high.p value to 1 for each feature.
If not, will scale to median of the high.p quantiles across the split.by
groups.}

\item{new.assay}{Name of new assay to save values to}
}
\value{
Returns a Seurat object with scaled values in the data slot of the
specified assay
}
\description{
Scale DSB assay to [0-1] per feature
}
\details{
Scales the expression values for each feature in the specified assay
from 0 to 1, with an optional split.by argument.
}
