% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeuratToAnnData.R
\name{SeuratToAnnData}
\alias{SeuratToAnnData}
\title{Save a Seurat object as an anndata h5ad file}
\usage{
SeuratToAnnData(
  object,
  outFile = NULL,
  main.assay = "RNA",
  main.layer = "counts",
  transfer.other.main.layers = NULL,
  transfer.other.assays.layers = NULL,
  transfer.graphs = TRUE
)
}
\arguments{
\item{object}{Seurat object}

\item{outFile}{Name of the target file to write the anndata object}

\item{main.assay}{Assay to transfer to AnnData as X}

\item{main.layer}{Layer to transfer to AnnData as X}

\item{transfer.other.main.layers}{Which other layers to transfer into AnnData}

\item{transfer.other.assays.layers}{Which other assay-layer pairs
to transfer into AnnData (since AnnData only takes one assay for now)}

\item{transfer.graphs}{Whether to transfer Graphs to AnnData .uns}
}
\value{
Writes a Seurat object to an .h5ad file after anndata conversion
}
\description{
Save a Seurat object as an anndata h5ad file
}
\references{
Code adapted from the awesome sceasy github!
}
