% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssignQuantiles.R
\name{AssignQuantiles}
\alias{AssignQuantiles}
\title{Assign cells to feature-specific quantiles.}
\usage{
AssignQuantiles(
  object,
  feature,
  assay = "ADT",
  slot = "data",
  split.by = "run_10x",
  quantile.probs = c(0, 0.25, 0.5, 0.75, 1)
)
}
\arguments{
\item{object}{Seurat object}

\item{feature}{Which feature to use for quantile calculation}

\item{assay}{Which assay the feature is located within}

\item{slot}{Which slot the to pull from within specified assay}

\item{split.by}{If not NULL, which metadata feature to split by before
calculating quantiles. Will then calculate quantiles within each split object}

\item{quantile.probs}{Which quantiles to split at}
}
\value{
Returns a Seurat object with quantile metadata added
}
\description{
Assign cells to feature-specific quantiles.
}
\details{
Determines the quantile of expression for each cell
for a specified feature in the Seurat object
}
