% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heatmaps.R
\name{DoStarHeatmap}
\alias{DoStarHeatmap}
\title{Heatmap of average expression values per group with stars}
\usage{
DoStarHeatmap(
  object,
  diff_exp_results = NULL,
  assay = "ADT",
  slot = "data",
  group.by = "seurat_clusters",
  max_zed = 3,
  plot_rownames = FALSE,
  label_features = TRUE,
  y_text_size = 10,
  scale_rows = TRUE,
  p_val_choice = 0.01,
  logFC = 0.4,
  plot_all = FALSE,
  plot_dendro = FALSE,
  subset_features = NULL,
  auc_choice = 0.6,
  cluster_cols = TRUE
)
}
\arguments{
\item{object}{Seurat object}
}
\value{
Returns a patchwork object of a heatmap and dendrogram if specified
}
\description{
Heatmap of average expression values per group with stars
}
\details{
TBD!!!
}
