% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlowPlot.R
\name{FlowPlot}
\alias{FlowPlot}
\title{Plot a metadata selection in a biaxial FeatureScatter plot}
\usage{
FlowPlot(
  object,
  feature1,
  feature2,
  assay = "ADT",
  slot = "data",
  pt.size = 1,
  noise.zero = FALSE,
  noise.zero.bound = -0.01,
  gg_bandwidth = 0.1,
  rasterize = FALSE
)
}
\arguments{
\item{object}{Seurat object}

\item{feature1}{The x-axis feature}

\item{feature2}{The y-axis feature}

\item{assay}{Which assay}

\item{slot}{Which slot in the assay, typically data}

\item{pt.size}{Size of geom_point}

\item{noise.zero}{Whether to simulate values from 0 to a negative value,
done for visual purposes ONLY and for kernel density estimation when
overlaying a density visualization. Will simulate values using a normal
distribution centered at 0, take absolute values, and then scale from
noise.zero.bound to 0.}

\item{noise.zero.bound}{The negative value to which to simulate zeros to.}

\item{rasterize}{Whether to rasterize the resulting plot.}
}
\value{
Returns a biaxial plot utilizing ggpointdensity::geom_pointdensity
}
\description{
Plot a metadata selection in a biaxial FeatureScatter plot
}
\details{
Scales the expression values for each feature in the specified assay
from 0 to 1, with an optional split.by argument
}
