% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunAUCell.R
\name{RunAUCell}
\alias{RunAUCell}
\title{Run AUCell on a Seurat object}
\usage{
RunAUCell(
  object,
  assay = NULL,
  slot = "counts",
  genesets,
  ranking.save = FALSE,
  ranking.key = NULL,
  normAUC = TRUE,
  aucMaxRank = 0.05,
  verbose = TRUE,
  auc_assay_name = "AUC",
  ...
)
}
\arguments{
\item{object}{Seurat object}

\item{assay}{Assay to use for building rankings. Will use default assay if NULL.}

\item{slot}{Slot to use for building rankings}

\item{genesets}{A list of vectors of features for expression programs; each entry should be a vector of feature names. If the list of vectors is named, the resulting AUCell score for the expression program will be a row in the AUCell assay returned.}

\item{ranking.save}{If TRUE, will save a new Assay `rankings` into `object` with the rankings of each gene per cell.}

\item{ranking.key}{If not NULL, will pull Assay `rankings` from `object` rather than re-calculating rankings.}

\item{normAUC}{Whether to normalize the maximum possible AUC per geneset to 1}

\item{aucMaxRank}{Threshold to calculate the AUC (see details section below and details section of \code{\link[AUCell]{AUCell_calcAUC}})}

\item{verbose}{Boolean. TRUE to show progress messages, FALSE to hide progress messages}
}
\value{
Returns a Seurat object with the AUCell results stored as a \code{\link{Assay}} object within the Seurat object
}
\description{
Run AUCell on a Seurat object
}
\details{
(Copied verbatim from AUCell::AUCell_calcAUC) _In a simplified way, the AUC value represents the fraction of genes, within the top X genes in the ranking, that are included in the signature. The parameter 'aucMaxRank' allows to modify the number of genes (maximum ranking) that is used to perform this computation. By default, it is set to 5% of the total number of genes in the rankings. Common values may range from 1 to 20%._
}
\references{
Aibar et al. (2017) SCENIC: single-cell regulatory network inference and clustering. Nature Methods. doi: 10.1038/nmeth.4463
}
\seealso{
\code{\link[AUCell]{AUCell_buildRankings}}

\code{\link[AUCell]{AUCell_calcAUC}}
}
