% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScaleDSB.R
\name{ScaleDSB}
\alias{ScaleDSB}
\title{Scale DSB assay to a global max per feature}
\usage{
ScaleDSB(
  object,
  assay = "DSB",
  slot = "data",
  split.by = NULL,
  high.p = 0.999,
  scale.to.one = TRUE,
  new.assay = "sDSB"
)
}
\arguments{
\item{object}{Seurat object}

\item{assay}{Which assay, typically DSB}

\item{slot}{Which slot in the assay, typically data after DSB correction}

\item{split.by}{If not NULL, which metadata feature to split by  before
scaling.}

\item{high.p}{Which percentile to set as the global max per object}

\item{scale.to.one}{Whether to set global max per feature as 1 (helpful when
needing equally weighted features)}

\item{new.assay}{Name of new assay to save values to}
}
\value{
Returns a Seurat object with scaled values in the data slot of the
specified assay
}
\description{
Scale DSB assay to a global max per feature
}
\details{
Scales the expression values for each feature in the specified assay
from 0 to a global max, with an optional split.by argument. The 0 is assumed
to be the 0 values in the slot, in line with DSB's procedure. We only scale
the high percentile.
}
