% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heatmaps.R
\name{TrueAverageExpression}
\alias{TrueAverageExpression}
\title{Average feature expression across clustered samples in a Seurat object using fast sparse matrix methods}
\usage{
TrueAverageExpression(
  object,
  group.by = "seurat_clusters",
  group.by.delim = "_",
  assay = "RNA",
  slot = "data",
  verbose = TRUE
)
}
\arguments{
\item{object}{Seurat object}

\item{group.by}{Ident with sample clustering information (default is
seurat_clusters). Can also input a vector with multiple idents specified and
TrueAverageExpression will return specified idents concatenated into new
a new output ident with group.by.delim as the separator.}

\item{group.by.delim}{The delimiter used if group.by is a vector of
length > 1.}

\item{assay}{Assay to average (default is the active assay)}

\item{slot}{Slot to average (default is counts)}

\item{verbose}{Boolean or integer, show progress bar (default is TRUE)}
}
\value{
dense matrix of cluster averages
}
\description{
Average feature expression across clustered samples in a Seurat object using fast sparse matrix methods
}
\references{
credit to GitHub user @zdebruine on Seurat issues!
}
